refactor(guidedcli): absorb HTTPMethod helpers from internal/chaosengine/handler
guidedcli was the only consumer of HTTPMethod/HTTPStatusCode helpers.
Move httphelpers.go + test into the guidedcli package (package guidedcli)
and drop the handler import from builders.go and resolver.go.
