+++ aegislab/src/internal/chaosengine/guidedcli/httphelpers_test.go
@@ -1,4 +1,4 @@
-package handler
+package guidedcli
 
 import (
 	"testing"
